Create shared file utiity module
